icons added & WeatherAdapter.kt added & TimeUtil added & strings.xml added and modified & Models added & Glide added & BindingAdapter.kt generated &  ApiConstants.kt added

+++ app/src/main/java/com/dmd/weatherforecast/models/Hourly.kt
@@ -3,7 +3,7 @@ package com.dmd.weatherforecast.models
 data class Hourly(
     val clouds: Int,
     val dew_point: Double,
-    val dt: Int,
+    val dt: Double,
     val feels_like: Double,
     val humidity: Int,
     val pop: Double,
