+++ app/src/main/java/com/dmd/weatherforecast/models/Minutely.kt
@@ -1,6 +1,6 @@
 package com.dmd.weatherforecast.models
 
 data class Minutely(
-    val dt: Int,
-    val precipitation: Int
+    val dt: Double,
+    val precipitation: Double
 )
